FunYa se puede crear la cuenta wallet al usuario recien creado, traer transacciones ok, login ok y registro usuario ok

+++ app/src/main/java/araya/gonzalo/alkewallwt/model/AccountRequest.kt
@@ -2,7 +2,7 @@ package araya.gonzalo.alkewallwt.model
 
 data class AccountRequest (
     val creationDate : String,
-    val money : Double,
+    val money : Int,
     val isBlocked : Boolean,
     val userId : Int
 )
